+++ src/types.ts
@@ -168,3 +168,14 @@ export type GamepadAxisName =
   | 'Right_Horizontal'
   | 'Right_Vertical'
   ;
+
+export type EventHandler<T> = (event: T) => void;
+
+export type UseEventTarget<E, T> = {
+  addEventListener: (type: E, fn: EventHandler<T>) => void;
+  removeEventListener: (type: E, fn: EventHandler<T>) => void;
+};
+
+export type UseEventListenersResult<E, T> = UseEventTarget<E, T> & {
+  fireEvent: (type: E, payload: T) => void;
+};